Fix circular dependency which was causing issues in CRA
