+++ packages/kerosene/src/datetime/constants.ts
@@ -1,0 +1,36 @@
+export const SECOND = 1_000;
+
+export const MINUTE = 60 * SECOND;
+
+export const HOUR = 60 * MINUTE;
+
+export const DAY = 24 * HOUR;
+
+export const DAYS_PER_WEEK = 7;
+
+export const MONTHS_PER_YEAR = 12;
+
+export enum Month {
+  JANUARY = 0,
+  FEBRUARY,
+  MARCH,
+  APRIL,
+  MAY,
+  JUNE,
+  JULY,
+  AUGUST,
+  SEPTEMBER,
+  OCTOBER,
+  NOVEMBER,
+  DECEMBER,
+}
+
+export enum DayOfWeek {
+  SUNDAY = 0,
+  MONDAY,
+  TUESDAY,
+  WEDNESDAY,
+  THURSDAY,
+  FRIDAY,
+  SATURDAY,
+}
