+++ packages/kerosene/src/datetime/index.ts
@@ -1,40 +1,4 @@
-export const SECOND = 1_000;
-
-export const MINUTE = 60 * SECOND;
-
-export const HOUR = 60 * MINUTE;
-
-export const DAY = 24 * HOUR;
-
-export const DAYS_PER_WEEK = 7;
-
-export const MONTHS_PER_YEAR = 12;
-
-export enum Month {
-  JANUARY = 0,
-  FEBRUARY,
-  MARCH,
-  APRIL,
-  MAY,
-  JUNE,
-  JULY,
-  AUGUST,
-  SEPTEMBER,
-  OCTOBER,
-  NOVEMBER,
-  DECEMBER,
-}
-
-export enum DayOfWeek {
-  SUNDAY = 0,
-  MONDAY,
-  TUESDAY,
-  WEDNESDAY,
-  THURSDAY,
-  FRIDAY,
-  SATURDAY,
-}
-
+export * from "./constants";
 export { default as getCalendarWeeks } from "./getCalendarWeeks";
 import {
   CalendarWeek as _CalendarWeek,
